refactor: 整理程式碼
將網際網路協定相關程式碼整理進資料夾

issue #51
