+++ 網際網路協定/ch01_gethostbyname.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+# Foundations of Python Network Programming, Third Edition
+# https://github.com/brandon-rhodes/fopnp/blob/m/py3/chapter01/getname.py
+
+import socket
+
+if __name__ == '__main__':
+    hostname = 'www.python.org'
+    addr = socket.gethostbyname(hostname)
+    print('The IP address of {} is {}'.format(hostname, addr))
+
